ntb_testwork/requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.0.6
idna==3.3
rel==0.4.7
requests==2.28.1
sqlparse==0.4.2
urllib3==1.26.9
websocket-client==1.3.3
